Delete Trailing Period function added
Added a checkbox to the UI to delete trailing periods. Added the code to support this. When the button is selected, files with trailing periods will appear. Wither click replace on the right to delete the period, or add a new character to replace the period with in the override.